Remove unused file Rename zodSchemaHelper to zodIntentDetectionHelper Rename Typechat to TypeChat Use full prompt in TypeChatLanguageModel Comment on ts-ignored import
